backend/requirements.txt
# FastAPI 核心框架
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 数据库相关
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# 认证和安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# 配置管理
pydantic==2.5.0
pydantic-settings==2.1.0

# 任务队列
celery==5.3.4
redis==5.0.1

# Manticore 搜索引擎客户端
pymysql==1.1.0

# HTTP 客户端
httpx==0.25.2
aiofiles==23.2.1

# 日志和监控
python-json-logger==2.0.7

# 开发和测试工具
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0

# 其他工具
python-dateutil==2.8.2
email-validator==2.1.0
psutil==5.9.6